Remove lots of hardcoded sizes
This improves rendering with DPI scaling (like on Windows) and with
backends/themes with very different control dimensions (like wxGTK3).

This was crosschecked with the GUI-related commits at wangqr/Aegisub but
the actual changes are done from scratch.

Reduce the upper bounds for a few spin controls (like the layer and
resolution) since wxGTK3 will always set the minimum content size of
spin controls to fit its maximum value, which will make the spinner
buttons clip outside the control if it's sized to something smaller,
see
https://github.com/wxwidgets/wxwidgets/issues/18568